backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
gunicorn==20.1.0
sentence-transformers==2.2.2
torch==2.6.0
open_clip_torch==2.24.0
faiss-cpu==1.7.4
numpy==1.26.4
python-multipart==0.0.6
pydantic==1.10.9
scikit-learn==1.3.2
Pillow==10.3.0
